FinalDose: -Cleaning up code for readability. -Adding electron calculations and autoselection of MC histories -Removed SRS from the grid test (1 mm resolution was resulting in a 1 hour computation) -Check on electron beams to see if they are clinical prior to computing dose -Added the UserInterface Status guide for encouraging user patience.
BeamOperations:
-Check first segment only on a beam. If the jaws on the first segment
need to be adjusted, then do it for all segments. (This sped up the
calculation significantly).
-Added nice formatting to jaw change message
-Set the tolerance for a suitable existing DSP to 0.01 Gy
-Added an alternative means of finding a dsp:
Find all points greater than specified percentage of dmax (say 98%)
Compute the centroid of these points
Return coordinates for the DSP.
emc_calc_params now determines the number of histories based on the maximum
of the hard limit (UW: 500000) and that required to give an uncertainty of
0.5%

Signed-off-by: rabayliss <37907358+rabayliss@users.noreply.github.com>