UWBrainCNS: Updated minimum coverage goal to MD preference
BeamOperations:
-Debug of check existing limits on stx to perform the check even if
the optimization has already started

OptimizationOperations:
-Added the automatic setting of treatment margins on VMAT and SMLC
plans to prevent the overexposure of normal tissues inadequately
optimized. This will force the aperture to not be never be larger
than a specified margin around any target roi which exists in the
Objectives list
-Extended this function to co-optimized beamsets to identify the
target being treated by a beamset

-Cleaned up the optimize_plan function in preparation for parameterizing
it using an input file.

Signed-off-by: rabayliss <37907358+rabayliss@users.noreply.github.com>