create_goals.py *Changed wording of the critical log message to make the TPO selected clear
planning_structures.py
*Moved all major functions to the the StructureOperations.py module
for the purposes of calling this function without the dialog

StructureOperations.py
*Added logcrit to this module. See about eliminating it long term
through message passing as we'd like to keep critical logging at the
top-level function calls
*make_boolean_structure moved here from planning_structures.py
*make_wall moved here from planning_structures.py
*check_structure_exists:
I'd like to confirm that the supplied strucutre not only exists in the
ROI list but has contours on the current exam. This will assist in
adaptation where multiple CT's are being used
-Added parameter for raystation exam to this function
Behavior should be unchanged if exam is not supplied.
:If the structure_name in the list of rois in this case:
 If the user wants it overrwritten, delete it and return False
 If the user wants a check, return True
If the exam is supplied:
:If the structure_name in the list of rois in this case:
::If the structure_name for this exam has ROIGeometries that have contours
-Either delete the geometry or return true
....it may be worth breaking these checks up and breaking up the delete
and checking operations

UWHeadNeck.xml
-Added clinical objectives for all ROIS to this protocol and cleaned
up the limits on targets. Tested on several patients. Found need to reduce
penalties on ipsilateral glands so I broke these limits up by protocol

AutoPlan_HN_Curative.py
Weird name...
Have the script based loading of planning structures working.

Signed-off-by: rabayliss <37907358+rabayliss@users.noreply.github.com>