Objectives.py -Bug that did not allow current case to be loaded when this function is called from a prompt.
StructureOperations.py
Adding matching dialog for normal structures. It will prompt the user for a suffix in the event that the structure list contains read-only (approved structures). In the event of a structure that is read only, that needs to have its name changed, will have its geometry copied to the correct structure name with a suffix. The Dice coefficient, specificity, etc. is output from this operation to a log file.

exists_roi:
-function is changes to use a case-insensitive reg-exp to look for a structure name.

match dialog:
Function was changed to add exclusions on planning structures from the match list.

create_roi:
-Significant complexity added (unfortunately). Function now looks for a geometry to exist in the case. If it finds one, and is supplied with an examination, it checks the current examination. It changes for approved structures as well. It is supposed to delete the geometry (for overwrite) , and cannot, then it adds a suffix.

StructureMatching:
-Added an output file to start tracking the matching users are doing to roi lists. This will feed into an Alias tag in the TG-263 file.

Signed-off-by: rabayliss <37907358+rabayliss@users.noreply.github.com>