Testing complete up to export liscense.
DicomExport.py
FIX: RS10A moves the energy field to a property of the beam. The loaded plan list was blank for export because the energy
    property is now null.

FIX: The energy of each beam is now a string value, the dictionary keys on the energies function are set to be text as well now.

DicomFilters.xml:
BUILD: We have combined RS machines from TrueBeam_FFF to TrueBeam. Now the FFF beams are specified in the quality id.
FIX: In RS Quality ID is "6 FFF" while in aria it is "6FFF"

BUILD: Electron fields have been moved to a single machine. All TrueBeam_E have been mapped to TrueBeam

BUILD: TomoTherapy: added 477 filter.