fix(automated_plan_optimization.py): Error capturing of failed optimization
Still working on capturing a failure to optimize. RS reports these with an Invalid-Operation Exception (Sys-level) exception. This is proving annoying to capture. I am trying to throw an IO error now when this fails. Tested and worked in several common errors like -dose grid not set, no volume in optimized ROI.