feature(See Description): Enabling lock to jaw limits
AutoPlanOperations.py, BeamOperations.py, OptiomizationOperations.py,SetTreatJawsCurrent.py
Enabling users to lock the jaws for optimization at their current limits and optimize.
This will help get us around new tests that are occurring on block settings on warm starts.

Also added was a change to the reduce_mod
operation in optimization. This will
compare the objective value of each
warmstart to the last and if the value
has increased, the previous iteration
will be loaded and reduce_mod
unflagged for the remaining warmstarts.

Signed-off-by: rabayliss <rabayliss@wisc.edu>