:ambulance: fix(Tomo3D High Dose and Large Targets): Improve optimization performance
When large targets or high dose targets are used in Tomo3D, especially with blocking, the optimization can
go poorly because the time is limited and no
modulation is possible. I have been using the
reduce_modulation operation to improve the time but
have been running into situations where reducing the
time is degrading plan quality. This change allows for the
previous warm start to be stored and if the optimization
value (for fixed objective functions) increases on the next
warmstart, the previous result is loaded and the time
reduction/reduce modulation is turned off for the
remaining iterations.

Signed-off-by: rabayliss <rabayliss@wisc.edu>