feature(automated_plan_optimization.py, OptimizationOperations.py): Adding Custom treatment settings
Allowing users to choose from:
-no treatment settings
-custom treatment settings
-standard treatment settings

Using a custom dialog for VMAT/SMLC versus TomoTherapy planning

Including the "reduce time" in the treatment optimization.
This will reduce the optimization time as long as it does not impact objective value. If the objective value is increased, then reload the patient at the prior to the last optimization.

Added improved error handling to the optimization.

Signed-off-by: rabayliss <rabayliss@wisc.edu>