:sparkles: perf(PRDR_CalcSheet.py): Added sorting table by field order
In order to deploy clinically, I revalidated this script using 5 anonymized patients.
The beams are first sorted by field order for ease of use in copying beams.

Merged the fractional and plan pandas frame
creation into a single function for speed.
Getting rid of the row number display since it is kind of easy to confuse with beam number.

Header updates and comments.