fix(AutoDQA.py): Fixed a sign error in shifts
Raystation flips the sign of AP for the y coordinate in the scripting API (iso and display value differ by a sign).
Shifting was fixed for the clipboard message. The shifting now is interpretted as the movement of the phantom relative to isocenter, e.g. shift Y -10 mm shifts the phantom toward the floor by 10 mm.

Also corrected isocenter name to match convention.