🚀 Test_AutoPlan.py: Update machine name in AutoPlan parameters
This commit updates the machine name in AutoPlan parameters from 'HDA0477' to 'HDA0488'. The changes were made to the file Test_AutoPlan.py at lines 74, 94, and 181. Additionally, the import statement for 'copy' and 'logcrit' were removed from lines 43 and 56 respectively. Finally, the autoplan_parameters argument was added to the function call on line 181.

Signed-off-by: Adam Bayliiss <37907358+rabayliss@users.noreply.github.com>