:recycle: AutoPlan.py: Refactor code to add ability to ignore status updates during script execution
This diff refactors the code in AutoPlan.py to add the ability to ignore status updates during script execution by introducing a new parameter `ignore_status`. If `ignore_status` is set to `True`, the script will no longer provide status updates during execution. This can be useful if the user wants to execute the script in the background or wants to suppress the status updates. The diff also includes updates to the `script_steps` dictionary and `UserInterface.ScriptStatus` class to handle the new parameter.

Signed-off-by: Adam Bayliiss <37907358+rabayliss@users.noreply.github.com>