:bug: 🔧 FinalDose.py: Refactor code to improve readability and add comments
This commit refactors the `FinalDose.py` code to improve readability and add comments for better understanding of the code. It also updates the `plan` and `case` object types to improve consistency and readability. Additionally, it removes unnecessary imports and variables. It modifies the ROI export list to only contain ROIs with clinical goals and adds a pattern to include ROIs with the word "block." It also improves the way ROIs are included and excluded from the export list, removing the try-catch blocks and adding more meaningful logging messages. Finally, it updates `StructureOperations.py` to correctly handle ROI exclusion from export.

Signed-off-by: Adam Bayliiss <37907358+rabayliss@users.noreply.github.com>