🐛 V2_autoplan_tomo_vmat_tbi.py: Fix OTV generation bug
This commit fixes a bug in the `make_otv` function that caused an error while generating the optimization target volume. The function was expecting a `namedtuple`, but it was receiving a `PatientData` object instead. To fix this, I added a type hint for the function parameters and ensured that the `poi_name` parameter contained an integer at the end. I also made some minor improvements to the comments and variable names in the code.

- Add type hints for function parameters
- Ensure poi_name contains integer at the end
- Improve comments and variable names

Signed-off-by: Adam Bayliiss <37907358+rabayliss@users.noreply.github.com>