🔧 gui_physics_review.py: Added comment_box argument to get_review_gui_values function
This commit includes modifications to gui_physics_review.py file. The main change is the addition of a new argument, comment_box, to the get_review_gui_values function. The purpose of the argument is to hold a dictionary with the contents of the Comments frame. Other changes include fixing a grammatical error in a comment and adjusting the pixel height for a Multiline element. Additionally, there are two TODOs added to log the passing and failed tests. Finally, a bulleted list is used to summarize the new arguments for get_review_gui_values function and the tabs variable assignments.

Signed-off-by: Adam Bayliiss <37907358+rabayliss@users.noreply.github.com>